Final Touch
At that step, you can already ``dotcloud push`` and get a working app;
but you still need to display a custom HTML form to start it.
To make testing more convenient, we will do a minor change to the code,
so that going to the MobWrite app will show the editor if no valid POST
parameter is supplied.

After this change, you can (re-)push the code, and this time, when
you go to the app URL, you should see the edit form. Opening the app
from multiple tabs or browsers will allow you to edit the form concurrently
(like Etherpad or recent versions of Google Docs Text Editor).