wsgi.py File
The "python" service dispatches HTTP requests to a WSGI-compatible callable
which must be found at ``wsgi.application``. In other words, we need to
have a ``application`` function in the ``wsgi.py`` file.

The MobWrite gateway has already been "WSGI-fied", and contains a suitable
``application`` callable. However, the gateway code is located in the
``daemon`` directory, and expects to import other modules from this
directory. Therefore, we will add the ``daemon`` directory to the Python Path
(instead of doing e.g. ``from daemon.gateway import application``).