supervisord.conf File
The MobWrite "gateway" (the WSGI application that we enabled in the previous
step) actually talks to the MobWrite "daemon". We need to start this "daemon".
While we could use ``dotcloud ssh`` to log into the service and start it
manually, we will rather write a ``supervisord.conf`` file, as explained in
the `Background Processes Guide <http://docs.dotcloud.com/guides/daemons/>`_.